[Add] Implement “with” methods to append or remove
Implement “with” methods such as “without_leading_zeros” and “without_leading_zeros!” methods in order to append or remove specific characters from a string object.
