+++ lib/string_foundation/version.rb
@@ -2,5 +2,5 @@
 # LIB - STRING FOUNDATION - VERSION
 # ==============================================================================
 module StringFoundation
-  VERSION = '0.1.0'.freeze
+  VERSION = '0.1.0'
 end
